refactor(server): rename user-related files for consistency
