+++ apps/server/src/user/user.controller.ts
@@ -9,7 +9,7 @@ import {
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { UsersService } from './users.service';
+import { UsersService } from './user.service';
 
 @Controller('users')
 export class UsersController {
